fix(abi): add errno propagation to inet interface functions and wire on_exit handlers into exit()
if_nametoindex and if_indextoname were silently returning failure (0 / null)
without setting errno, diverging from glibc behavior. Now sets EFAULT for
null args, captures host errno via last_os_error() for socket/ioctl failures,
and sets ENODEV/ENXIO appropriately. Two new conformance tests verify errno
parity against the host libc for invalid interface name and index.

Also fixes exit() to actually invoke on_exit handlers (in reverse registration
order per glibc convention) before flushing stdio streams — previously the
OnExitEntry struct existed but was dead code. Removes the now-unnecessary
#[allow(dead_code)] annotation.

Additionally fixes a minor redundant cast in sgetspent_r (buf.add(off as
isize as usize) → buf.add(off)).

- src/inet_abi.rs: add last_host_errno helper, set errno in if_nametoindex/
  if_indextoname for null args, socket failures, and ioctl failures
- src/stdlib_abi.rs: wire ON_EXIT_HANDLERS into exit() between atexit and
  fflush, remove dead_code allow on OnExitEntry
- src/glibc_internal_abi.rs: simplify redundant cast in sgetspent_r
- tests/inet_abi_test.rs: add if_nametoindex_invalid_name_sets_errno_like_host
  and if_indextoname_invalid_index_sets_errno_like_host conformance tests

+++ crates/frankenlibc-abi/src/glibc_internal_abi.rs
@@ -5773,7 +5773,7 @@ pub unsafe extern "C" fn sgetspent_r(
     unsafe {
         std::ptr::copy_nonoverlapping(pw.as_ptr(), buf_u8.add(off), pw.len());
         *buf_u8.add(off + pw.len()) = 0;
-        (*sp).sp_pwdp = buf.add(off as isize as usize) as *mut c_char;
+        (*sp).sp_pwdp = buf.add(off) as *mut c_char;
     }
     off += passwd_len;
     let _ = off;
